Don't depend on libvirt-daemon-kvm monolith.
libguestfs usually needs qemu.  However it only requires an emulator
for the same architecture, not for all architectures.
libvirt-daemon-kvm pulls in qemu which pulls in emulators for all
architectures, as well as a bunch of other stuff we don't need at all
like network interface support and nwfilter.

There are no Fedora TCG-only arches, so drop the conditional section.